% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvToDataframe.R
\name{csvToDataframe}
\alias{csvToDataframe}
\title{csvToDataframe}
\usage{
csvToDataframe(s3URL)
}
\arguments{
\item{s3URL}{This is the S3 URL to the .csv file.}
}
\value{
A dataframe made from .csv.
}
\description{
This function returns a dataframe from a .csv file stored on AWS S3.
}
\examples{
csvToDataframe("s3://cloudsimplifier-examples/example.csv")
}
