% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileData.R
\name{getFileData}
\alias{getFileData}
\title{getFileData}
\usage{
getFileData(filename, format, bucket)
}
\arguments{
\item{filename}{This is the name of the file with extension.}

\item{format}{This is the format of the file. Currently supports: .csv, .json}

\item{bucket}{This is the name of the S3 Bucket.}
}
\value{
A dataframe
}
\description{
This function returns a dataframe from a file stored on an AWS S3 Bucket.
}
\examples{
getFileData("example.csv", "csv", "cloudsimplifier-examples")
}
