% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilenames.R
\name{getFilenames}
\alias{getFilenames}
\title{getFilenames}
\usage{
getFilenames(bucket)
}
\arguments{
\item{bucket}{This is the bucketname from which to return files. This function assumes there is no hierarchy and won't dive into folders.}
}
\value{
filenames A simple list of files in the S3 Bucket
}
\description{
This function returns a list of filenames in a specified S3 Bucket. This will not work until you set environment variables using "setEnvironment"
}
