% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonToDataframe.R
\name{jsonToDataframe}
\alias{jsonToDataframe}
\title{jsonToDataframe}
\usage{
jsonToDataframe(s3URL)
}
\arguments{
\item{s3URL}{This is the S3 URL to the json file.}
}
\value{
A dataframe made from json. If the json is invalid, it prints an error message and returns the json string for debugging or other exploration.
}
\description{
This function returns a dataframe from a json file stored on AWS S3.
}
\examples{
jsonToDataframe("s3://cloudsimplifier-examples/example.json")
}
