% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setEnvironment.R
\name{setEnvironmnet}
\alias{setEnvironmnet}
\title{setEnvironment}
\usage{
setEnvironmnet(key, secret, region)
}
\arguments{
\item{key}{This goes into the AWS_ACCESS_KEY_ID}

\item{secret}{This goes into the AWS_SECRET_ACCESS_KEY}

\item{region}{This goes into the AWS_DEFAULT_REGION}
}
\description{
This function sets the environmnet variables for the AWS environment
}
